Tweaked and added refactoring for main, test and docs
Some links are accidentally refactored in docs in the previous
commit. Main and test have un-refactored portions.

Let's undo all erroneous refactoring in the previous commit in docs
and add more refactoring to main and test.

Some links are needed to extract code coverage related information
from the addressbook website.

UserGuide.md has links that need to change later on. This commit
will refactor the class diagrams incompletely and is done so due to
undecidedness. The class diagrams will be refactored later if need
be.