Enhance UI mockup
Ui mockup currently does not reflect the future plans as well as
the current features of the app.

Let's modify the ingredient list visual to a more list-like visual
with the units specified along with the ingredient name and
quantity. Let's also move the command result visual to the bottom
right of the app Ui mockup.

The ingredient list visual is modified to represent the unit
attribute according to the ingredient class. The command result is
moved and expanded to accommodate the viewing of large recipes, a
feature coming up in future iterations.