Fix bugs for the view recipe feature
Implementation will cause invocation error if ViewRecipeCommand is
called on a uuid belonging to a recipe that has been deleted. This
is due to improper invalid uuid checks in the associated parser
class. Invalid uuid exception handling is also in the parse
method in the ViewRecipeCommandParser class and not in the execute
method in ViewRecipeCommand.

The invocation error is simply a bug that needs to be remedied. The
handling of invalid uuid should not be handled by the parser since
it is technically not a parse exception.

Let's fix the bug associated with viewing deleted recipes, and shift
invalid uuid checks to the ViewRecipeCommand class.