Add search command
There is no way to look for a recipe based on a
specific ingredient.

Let's add a command that will do just that

The search command will allow users to search for any recipe
that has the corresponding ingredient name.

Testing has also been done on the SearchCommand class and the
SearchCommandParser class