Clean code design
The parser and commands use the Index class as a way to encapsulate
the Unique Identifier(uuid) of each recipe or use an integer value
to represent the uuid.

The Index class seems to be a unneccessary representation of the
uuid since there is already a UniqueId class present and many of
the methods in Index do not seem customised for uuid. Representing
the uuid as an integer also does not seem to be the way to go as
this makes the code less robust and extensible in the future.

Let's replace the parsing and execution methods to incorporate the
uuid as a UniqueId object instead of an Index object or an integer.

The delete tests are commented out for the time being.