Implement bugfixes for UseCommand
UseCommand feature allows negative input to be keyed in by user. The
parse excpetions raised in parsing the quantity of ingredient refers
to the quantity as "amount" instead.

Let's raise a parse exception when the user inputs a negative
quantity, and change the parseAmount method in ParserUtil class to
parseQuantity instead. The method will take in an extra unit
parameter, and output a Quantity object instead of a double value.

The method was changed to make it more specific towards parsing
only quantity since our Recipe Book never needs to parse a raw
number. This also allows the error messages in the parse exception
to be more appropriate.

The relevant tests for negative values are also added.